+++ Assets/02.Scripts/Player/IDamagable.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public interface IDamagable
 {
-    void TakeDamage(float damage);
+    void TakeDamage(int damage);
 
     void Die();
 }
